Create component FetchCharacters
